requirements.txt
anthropic>=0.25.0
streamlit>=1.35.0
python-dotenv>=1.0.0
pandas>=2.0.0
Pillow>=10.0.0
requests>=2.31.0
